MAJOR REFACTOR: Use separate Recorder Worker
Previously, saving (and returning) of the TileProcessWorker is handled
by the workers. That results in locks and inefficiencies due to having
to synchronize some non-trivial structs (e.g.m dict) between the
processes.

This major change makes the TileProcessorWorkers *only* do the
DominantColor calculation and hands over result of calculation to a
process dedicated to recording the results (and saving).

The Recorder process is further optimized by:
- Enforce a "minimum distance" between saves
- Do not return results via synchronized structs, but through a much
  simpler and performant queue
- ... and return results only when asked.