Drain coordfail_q as well
It's possible coordfail_q to contain items if for some reason Recorder
doesn't manage to invoke its _flush() method.

Or if Recorder dies and in its last gasps managed to push something
into coordfail_q.

With this while loop, we ensure that coordfail_q is both drained *and*
recorded.

Also added some type hints to the queue objects to prevent unpacking
confusion.