Define the keys inside DominantColors
This way we reduce the intimate coupling between MosaicMap and
DominantColors: We no longer have to delve in to find the exact keys,
we will just trust DominantColors to provide the applicable keys for
the subtiling strategy.

Also, make the key collections immutable (tuples), and in MosaicMap
accept as Sequence.