Add 'quiet' property to WorkTeam
This property gets propagated to all Workers in the team.

We rename the `quiet` arg of start() to prevent confusion (because that
arg only silences the startup process, not the workers themselves).

In addition, we also add a `quiet` arg to disband() so the workers can
be selectively quieted (or not) during disband(). Also forwarded to the
pre_disband() and post_disband() methods.

Finally, we add more `if not quiet` tests to the Workers themselves.