MAJOR REFACTOR (see desc)
* MOVE all drawing logic to roadmapper_v3.draw
* This greatly simplifies model to really just model with no drawing smarts
* And we managed to remove the ugly callbacks (sorry Sir pepo; they *are* ugly!)
* This unexpectedly also lead to greatly simplified imports for some other modules

A very important change is moving Point the other way around: From .draw to .model

This is mainly to prevent circular import. But it turned out to be the right thing
to do, because Point is after all another model, it does not know how to do things
beyond its own scope. And ultimately resulted in that modules that do not concern
themselves with how to draw, do not need to import anything from .draw

Nice!