Add a Collector Worker
After some thoughts, this is better, I think.

The Collector Worker acts as the buffer between the Calc
Workers and the Mosaic Maker. The RLock before accessing
the Managed Dict, though not strictly necessary, ensures
that writes are done before Mosaic Maker works. Had this
lock been in the Calc Workers, we might be wasting CPU
cycles just to wait for the lock.