Change break condition
Previously it immediately breaks when None is received, this results in still
in-flight tasks to fail.

Happened only when the job list is exhausted so we missed that.

Now if None is specified it also checks for outstanding jobs before breaking
out of the loop.

Also, we move the detection after the asyncio.wait logic so we actually
let the in-flight tasks to complete.